pr-fix: split az func test classes
