+++ src/Arcus.WebApi.Tests.Unit/Logging/Fixture/AzureFunctions/TestHttpResponseData.cs
@@ -1,0 +1,19 @@
+﻿using System.IO;
+using System.Net;
+using Microsoft.Azure.Functions.Worker;
+using Microsoft.Azure.Functions.Worker.Http;
+
+namespace Arcus.WebApi.Tests.Unit.Logging.Fixture.AzureFunctions
+{
+    public class TestHttpResponseData : HttpResponseData
+    {
+        public TestHttpResponseData(FunctionContext functionContext) : base(functionContext)
+        {
+        }
+
+        public override HttpStatusCode StatusCode { get; set; }
+        public override HttpHeadersCollection Headers { get; set; } = new HttpHeadersCollection();
+        public override Stream Body { get; set; } = new MemoryStream();
+        public override HttpCookies Cookies { get; }
+    }
+}
